pyproject.toml
[tool.poetry]
name = "crypto-trading"
version = "0.1.0"
description = "trading infrastructure to algorithmically trade crypto currencies"
authors = ["David Diranko"]
readme = "README.md"
classifiers = [
        "Programming Language :: Python :: 3.8"]
packages = [{include = "crypto_trading"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
pandas = "==1.5.0"
boto3 = "==1.24.86"
pybit = "==2.4.0"
python-binance = "==1.0.16"
python-dotenv = "==0.20.0"
scikit-learn = "==1.2.0"
trading-ig = "==0.0.20"
aiohttp = "==3.7.4.post0"
pyaml = "==6.0.1"

[tool.poetry.dev-dependencies]
isort = "==5.10.1"
yapf = "==0.32.0"
toml = "==0.10.2"
mypy = "==0.982"
coverage = "==6.5.0"
jupyter = "==1.0.0"
tqdm = "==4.64.1"
openpyxl = "==3.0.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"